requirements.txt
fastapi==0.73.0
google-cloud-storage==2.1.0
gunicorn==20.1.0
matplotlib==3.5.1
numpy==1.22.1
Pillow==9.0.0
python-multipart==0.0.5
torch==1.10.1
torchvision==0.11.2
uvicorn==0.17.0

